# indexer/services/comlink/public/swagger.json
{
  "components": {
...
    "schemas": {
...
      "IsoString": {
        "type": "string"
      },
...
      "OrderSide": {
        "enum": [
          "BUY",
          "SELL"
        ],
        "type": "string"
      },
      "Liquidity": {
        "enum": [
          "TAKER",
          "MAKER"
        ],
        "type": "string"
      },
      "FillType": {
        "enum": [
          "LIMIT",
          "LIQUIDATED",
          "LIQUIDATION",
          "DELEVERAGED",
          "OFFSETTING"
        ],
        "type": "string"
      },
      "FillResponseObject": {
        "properties": {
          "id": {
            "type": "string"
          },
          "side": {
            "$ref": "#/components/schemas/OrderSide"
          },
          "liquidity": {
            "$ref": "#/components/schemas/Liquidity"
          },
          "type": {
            "$ref": "#/components/schemas/FillType"
          },
          "market": {
            "type": "string"
          },
          "price": {
            "type": "string"
          },
          "size": {
            "type": "string"
          },
          "fee": {
            "type": "string"
          },
          "createdAt": {
            "$ref": "#/components/schemas/IsoString"
          },
          "createdAtHeight": {
            "type": "string"
          },
          "orderId": {
            "type": "string"
          },
          "clientMetadata": {
            "type": "string"
          },
          "subaccountNumber": {
            "type": "number",
            "format": "double"
          }
        },
        "required": [
          "id",
          "side",
          "liquidity",
          "type",
          "market",
          "price",
          "size",
          "fee",
          "createdAt",
          "createdAtHeight",
          "subaccountNumber"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "FillResponse": {
        "properties": {
          "fills": {
            "items": {
              "$ref": "#/components/schemas/FillResponseObject"
            },
            "type": "array"
          }
        },
        "required": [
          "fills"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {}
  },
  "info": {
    "title": "Klyra Indexer API",
    "version": "v1.0.0",
    "description": "Scroll down for code samples, example requests and responses.\nBase URLs:\n* For **Testnet**, use <a href=\"https://klyra-testnet.imperator.co/v4\">https://klyra-testnet.imperator.co/v4</a>\nNote: Messages on Indexer WebSocket feeds are typically more recent than data fetched via Indexer's REST API, because the latter is backed by read replicas of the databases that feed the former. Ordinarily this difference is minimal (less than a second), but it might become prolonged under load."
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/fills": {
      "get": {
        "operationId": "GetFills",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FillResponse"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "address",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "subaccountNumber",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "market",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "createdBeforeOrAtHeight",
            "required": false,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "createdBeforeOrAt",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/IsoString"
            }
          }
        ]
      }
    },
    "/fills/parentSubaccount": {
      "get": {
        "operationId": "GetFillsForParentSubaccount",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FillResponse"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "address",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "parentSubaccountNumber",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "market",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "createdBeforeOrAtHeight",
            "required": false,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "createdBeforeOrAt",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/IsoString"
            }
          }
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://klyra-testnet.imperator.co/v4",
      "description": "Testnet"
    }
  ]
}